Feature:LoginMoudule - Added basic structure of LoginButton and LoginDialogue
Add a basic responsive login button which can call out the LoginDialogue, where the dialogue could catch what user input and save it into a userSubmit object, which is ready to be submitted by the request. Some of the mentioned changes is related to last two commits, if code review needed please check last to commits.